requirements.txt
django
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
dj-config-url
python-decouple
drf-yasg
pillow
django-filter
celery[redis]
loguru
coverage
model_bakery
